# docs/swagger.yaml
basePath: /api
definitions:
...
  questions.CreateQuestionRequest:
    properties:
      correct_answer:
        type: string
      display_order:
        minimum: 1
        type: integer
      explanation:
        type: string
      points:
        minimum: 1
        type: integer
      question_data:
        type: object
      question_text:
        type: string
      question_type:
        type: string
      section_id:
        type: string
    required:
    - display_order
    - question_text
    - question_type
    - section_id
    type: object
  questions.UpdateQuestionRequest:
    properties:
      correct_answer:
        type: string
      display_order:
        minimum: 1
        type: integer
      explanation:
        type: string
      is_active:
        type: boolean
      points:
        minimum: 1
        type: integer
      question_data:
        type: object
      question_text:
        type: string
    required:
    - display_order
    - question_text
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Backend API for English proficiency testing platform.
  title: Go English Test Platform API
  version: "1.0"
paths:
...
  /questions:
    get:
      description: Get list of questions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all questions
      tags:
      - Questions
    post:
      consumes:
      - application/json
      description: Create a new question
      parameters:
      - description: Create question request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/questions.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create question
      tags:
      - Questions
  /questions/{id}:
    delete:
      description: Delete question by ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete question
      tags:
      - Questions
    get:
      description: Get question detail by ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get question by ID
      tags:
      - Questions
    put:
      consumes:
      - application/json
      description: Update existing question
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Update question request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/questions.UpdateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update question
      tags:
      - Questions
  /questions/by-section:
    get:
      description: Get questions by section ID
      parameters:
      - description: Section ID
        in: query
        name: section_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get questions by section ID
      tags:
      - Questions
  /results:
    get:
      description: Get all results for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get my results
      tags:
      - Results
  /results/{id}:
    get:
      description: Get result detail by ID
      parameters:
      - description: Result ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get result by ID
      tags:
      - Results
  /results/by-attempt:
    get:
      description: Get result detail by attempt ID
      parameters:
      - description: Attempt ID
        in: query
        name: attempt_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get result by attempt ID
      tags:
      - Results
  /roles:
    get:
      description: Get list of available roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Roles
  /roles/{id}:
    get:
      description: Get role detail by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Roles
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"